requirements.txt
# Sanic
sanic
sanic-ext
oss2
python-dotenv
